Reverted back to the 'Libft' version of the function, with argument 'start' of type 'unsigned int'

+++ lib/libft/ft_substr.c
@@ -6,13 +6,13 @@
 /*   By: rtorrent <rtorrent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/07 12:43:27 by rtorrent          #+#    #+#             */
-/*   Updated: 2023/07/04 17:50:22 by rtorrent         ###   ########.fr       */
+/*   Updated: 2024/08/14 21:40:13 by rtorrent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(const char *s, size_t start, size_t len)
+char	*ft_substr(const char *s, unsigned int start, size_t len)
 {
 	size_t	size;
 	char	*p;
